Split file up
